# sdk/python/contract/openapi.json
{
  "components": {
    "schemas": {
...
      "Edge": {
        "properties": {
          "agent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Id"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "properties": {
            "title": "Properties",
            "type": "object"
          },
          "relation": {
            "title": "Relation",
            "type": "string"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "target": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target"
          },
          "target_id": {
            "title": "Target Id",
            "type": "string"
          },
          "utility_weight": {
            "default": 1.0,
            "description": "Synapse strength (V2 Cognitive Graph). Controls edge width in 3D view.",
            "title": "Utility Weight",
            "type": "number"
          }
        },
        "required": [
          "source_id",
          "target_id",
          "relation"
        ],
        "title": "Edge",
        "type": "object"
      },
...
      "GraphPayload": {
        "properties": {
          "edges": {
            "items": {
              "$ref": "#/components/schemas/Edge"
            },
            "title": "Edges",
            "type": "array"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/Node"
            },
            "title": "Nodes",
            "type": "array"
          }
        },
        "title": "GraphPayload",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IngestQueuedResponse": {
        "properties": {
          "agent_id": {
            "title": "Agent Id",
            "type": "string"
          },
          "message_id": {
            "title": "Message Id",
            "type": "string"
          },
          "status": {
            "default": "queued",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "message_id",
          "agent_id"
        ],
        "title": "IngestQueuedResponse",
        "type": "object"
      },
      "IngestRequest": {
        "properties": {
          "agent_id": {
            "default": "default",
            "title": "Agent Id",
            "type": "string"
          },
          "metadata": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "text": {
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "IngestRequest",
        "type": "object"
      },
...
      "Node": {
        "properties": {
          "agent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Id"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "embedding": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "properties": {
            "title": "Properties",
            "type": "object"
          },
          "type": {
            "default": "OTHER",
            "title": "Type",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "label"
        ],
        "title": "Node",
        "type": "object"
      },
      "QueryRequest": {
        "properties": {
          "agent_id": {
            "default": "default",
            "title": "Agent Id",
            "type": "string"
          },
          "question": {
            "title": "Question",
            "type": "string"
          },
          "session_id": {
            "default": "",
            "description": "Client-supplied session identifier for grouping related interaction turns. Used by episodic memory (interaction_memory=True) to link related queries. A UUID is generated server-side when omitted.",
            "title": "Session Id",
            "type": "string"
          },
          "top_k": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Number of seed nodes retrieved by vector search. Downstream context limits scale proportionally (context_limit = ceil(top_k * 2.5)). Omit to use the server default (top_k=8). Recommended: 8 for small graphs (<1k nodes), 20 for medium, 50 for large (50k+ nodes).",
            "title": "Top K"
          }
        },
        "required": [
          "question"
        ],
        "title": "QueryRequest",
        "type": "object"
      },
      "QueryResult": {
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "confidence_score": {
            "default": 1.0,
            "description": "Verifier confidence score from the final iteration.",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence Score",
            "type": "number"
          },
          "cypher_used": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cypher Used"
          },
          "iterations_used": {
            "default": 1,
            "description": "Number of retrieve-verify iterations executed before synthesis.",
            "title": "Iterations Used",
            "type": "integer"
          },
          "question": {
            "title": "Question",
            "type": "string"
          },
          "re_query_happened": {
            "default": false,
            "description": "True when at least one re-query was issued due to insufficient context.",
            "title": "Re Query Happened",
            "type": "boolean"
          },
          "subgraph": {
            "$ref": "#/components/schemas/GraphPayload"
          },
          "verifier_feedback": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Accumulated missing-information categories reported by the verifier across all iterations. None when the first pass was sufficient.",
            "title": "Verifier Feedback"
          }
        },
        "required": [
          "question",
          "answer",
          "subgraph"
        ],
        "title": "QueryResult",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "SpAIder \u2014 A multi-agent semantic knowledge graph platform powered by LLMs and Neo4j. Ingest unstructured text, build a persistent knowledge graph, query it in natural language, and synthesise fine-tuning datasets for downstream models.",
    "title": "SpAIder",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/analytics/ingest": {
      "get": {
        "description": "Hourly ingest volume, node/edge counts, and avg latency for the last N days.",
        "operationId": "ingest_timeseries_api_v1_analytics_ingest_get",
        "parameters": [
          {
            "in": "query",
            "name": "agent_id",
            "required": false,
            "schema": {
              "default": "default",
              "title": "Agent Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 7,
              "maximum": 90,
              "minimum": 1,
              "title": "Days",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest Timeseries",
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/v1/analytics/overview": {
      "get": {
        "description": "Aggregate totals: ingests, nodes created, queries, latencies.",
        "operationId": "overview_api_v1_analytics_overview_get",
        "parameters": [
          {
            "in": "query",
            "name": "agent_id",
            "required": false,
            "schema": {
              "default": "default",
              "title": "Agent Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Overview",
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/v1/analytics/query": {
      "get": {
        "description": "Hourly query volume, avg latency, and nodes returned for the last N days.",
        "operationId": "query_timeseries_api_v1_analytics_query_get",
        "parameters": [
          {
            "in": "query",
            "name": "agent_id",
            "required": false,
            "schema": {
              "default": "default",
              "title": "Agent Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 7,
              "maximum": 90,
              "minimum": 1,
              "title": "Days",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Query Timeseries",
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/v1/analytics/top-agents": {
      "get": {
        "description": "Most active agents by ingest count.",
        "operationId": "top_agents_api_v1_analytics_top_agents_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Top Agents",
        "tags": [
          "analytics"
        ]
      }
    },
...
    "/api/v1/ingest": {
      "post": {
        "description": "Async ingest: serialise the payload and push to the Kafka topic.\nThe compressor worker will consume and write to Neo4j asynchronously.",
        "operationId": "ingest_text_async_api_v1_ingest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestQueuedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest Text Async",
        "tags": [
          "ingest"
        ]
      }
    },
...
    "/api/v1/query": {
      "post": {
        "description": "Natural-language query over the knowledge graph.\nTranslates the question to Cypher using an LLM, executes it, and synthesises\na human-readable answer with supporting subgraph context.\n\n**Episodic Memory:** When the agent has ``interaction_memory=True``, the\nquery/response pair is recorded asynchronously as an ``InteractionNode``\nin Neo4j (fire-and-forget via BackgroundTasks \u2014 zero added latency).",
        "operationId": "query_nl_api_v1_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Query Nl",
        "tags": [
          "query"
        ]
      }
    },
...
  }
}